backend/requirements.txt
fastapi
netmiko
netdev
ntc-templates
uvicorn
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]